feat: ui editor & new 30ft skins

+++ RWLib/Graphics/RandomSkin.cs
@@ -12,6 +12,9 @@ namespace RWLib.Graphics
     {
         [JsonPropertyName("id")]
         public string Id { get; set; } = "";
+        [JsonPropertyName("kind")]
+        // Groups with the same kind can share eachother skins when filling empty spots
+        public string? Kind { get; set; } = null;
 
         [JsonPropertyName("randomSkins")]
         public List<RandomSkin> RandomSkins { get; set; } = new List<RandomSkin>();
@@ -60,8 +63,19 @@ namespace RWLib.Graphics
             public string Texture { get; set; } = "";
         }
 
-        public void OrderSkins()
+        public void FillAndOrderSkins(List<RandomSkinGroup> relatedGroups)
         {
+            var relatedSkins = relatedGroups
+                .Select(x => x.RandomSkins)
+                .SelectMany(x => x)
+                .Where(x => x.Composition == this.Composition && x.Id == this.Id)
+                .SelectMany(x => x.Skins)
+                .OrderBy(x => {
+                    var alreadyHas = this.Skins.Any(s => s.Texture == x.Texture);
+                    return alreadyHas ? 1 : 0; // prioritize skins that are not already in the list
+                })
+                .ToList();
+
             var skins = Skins.OrderByDescending(x => x.Rarity).ToList();
 
             var duplicates = skins.GroupBy(x => x.Texture)
@@ -76,7 +90,17 @@ namespace RWLib.Graphics
             while (skins.Count < FullSkinsAmount)
             {
                 Console.WriteLine(this.Id + " Composition is not fully filled. The remaining space will be filled with duplicates.");
-                skins.AddRange(skins.ToArray());
+                if (relatedSkins.Count == 0 && skins.Count == 0)
+                {
+                    throw new InvalidDataException("No skins available to fill the composition");
+                }
+                if (relatedSkins.Count > 0)
+                {
+                    Console.WriteLine("Filling from related skins: " + relatedSkins.Count + " available");
+                }
+                var dups = skins.ToArray();
+                skins.AddRange(relatedSkins);
+                skins.AddRange(dups);
                 if (skins.Count > FullSkinsAmount)
                 {
                     skins.RemoveRange(FullSkinsAmount, skins.Count - FullSkinsAmount);
@@ -89,6 +113,7 @@ namespace RWLib.Graphics
                 throw new InvalidDataException("More skins found than the maximum allowed");
             }
 
+            skins = skins.OrderByDescending(x => x.Rarity).ToList();
             this.Skins = skins;
         }
     }
